+++ scanpipe/pipelines/collect_symbols_ctags.py
@@ -26,8 +26,8 @@ from scanpipe.pipes.input import copy_inputs
 from scanpipe import pipes
 
 
-class CollectSymbols(Pipeline):
-    """Collect symbols from codebase files and keep them in extra data field."""
+class CollectSymbolsCtags(Pipeline):
+    """Collect source symbols with Ctags."""
 
     download_inputs = False
     is_addon = True
@@ -58,4 +58,4 @@ class CollectSymbols(Pipeline):
         Collect symbols from codebase files using Ctags and store
         them in the extra data field.
         """
-        symbols.collect_and_store_resource_symbols(self.project, self.log)
+        symbols.collect_and_store_resource_symbols_ctags(self.project, self.log)