+++ scanpipe/pipelines/deploy_to_develop.py
@@ -70,6 +70,8 @@ class DeployToDevelop(Pipeline):
             cls.map_java_to_class,
             cls.map_jar_to_source,
             cls.map_javascript,
+            cls.map_elf,
+            cls.map_go,
             cls.match_directories_to_purldb,
             cls.match_resources_to_purldb,
             cls.map_javascript_post_purldb_match,
@@ -188,6 +190,16 @@ class DeployToDevelop(Pipeline):
         to its source.
         """
         d2d.map_javascript(project=self.project, logger=self.log)
+
+    @group("Elf")
+    def map_elf(self):
+        """Map ELF binaries to their sources."""
+        d2d.map_elfs(project=self.project, logger=self.log)
+
+    @group("Go")
+    def map_go(self):
+        """Map Go binaries to their sources."""
+        d2d.map_go_paths(project=self.project, logger=self.log)
 
     def match_directories_to_purldb(self):
         """Match selected directories in PurlDB."""