+++ scanpipe/pipes/resolve.py
@@ -42,6 +42,16 @@ Resolve packages from manifest, lockfile, and SBOM.
 """
 
 
+def resolve_manifest_resources(resource, package_registry):
+    """Get package data from resource."""
+    packages = get_packages_from_manifest(resource.location, package_registry) or []
+
+    for package_data in packages:
+        package_data["codebase_resources"] = [resource]
+
+    return packages
+
+
 def get_packages(project, package_registry, manifest_resources, model=None):
     """
     Get package data from package manifests/lockfiles/SBOMs or
@@ -51,17 +61,17 @@ def get_packages(project, package_registry, manifest_resources, model=None):
 
     if not manifest_resources.exists():
         project.add_warning(
-            description="No resources found with package data",
+            description="No resources containing package data found in codebase.",
             model=model,
         )
-        return
+        return []
 
     for resource in manifest_resources:
-        if packages := get_packages_from_manifest(resource.location, package_registry):
+        if packages := resolve_manifest_resources(resource, package_registry):
             resolved_packages.extend(packages)
         else:
             project.add_error(
-                description="No packages could be resolved for",
+                description="No packages could be resolved",
                 model=model,
                 resource=resource,
             )
@@ -80,7 +90,8 @@ def create_packages_and_dependencies(project, packages, resolved=False):
     for package_data in packages:
         package_data = set_license_expression(package_data)
         dependencies = package_data.pop("dependencies", [])
-        update_or_create_package(project, package_data)
+        codebase_resources = package_data.pop("codebase_resources", [])
+        update_or_create_package(project, package_data, codebase_resources)
 
         for dependency_data in dependencies:
             if resolved:
@@ -264,17 +275,17 @@ def get_default_package_type(input_location):
         if handler.is_datafile(input_location):
             return handler.default_package_type
 
-        if input_location.endswith((".spdx", ".spdx.json")):
-            return "spdx"
+    if input_location.endswith((".spdx", ".spdx.json")):
+        return "spdx"
 
-        if input_location.endswith((".bom.json", ".cdx.json")):
+    if input_location.endswith(("bom.json", ".cdx.json", "bom.xml", ".cdx.xml")):
+        return "cyclonedx"
+
+    if input_location.endswith((".json", ".xml")):
+        if cyclonedx.is_cyclonedx_bom(input_location):
             return "cyclonedx"
-
-        if input_location.endswith(".json"):
-            if cyclonedx.is_cyclonedx_bom(input_location):
-                return "cyclonedx"
-            if spdx.is_spdx_document(input_location):
-                return "spdx"
+        if spdx.is_spdx_document(input_location):
+            return "spdx"
 
 
 # Mapping between `default_package_type` its related resolver functions